# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portal Trigo",
    "description": "Backend do Portal Trigo.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/protheus/status": {
      "get": {
        "tags": [
          "protheus"
        ],
        "summary": "Protheus Status",
        "operationId": "protheus_status_api_protheus_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtheusStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/protheus/test-connection": {
      "post": {
        "tags": [
          "protheus"
        ],
        "summary": "Protheus Testar",
        "description": "Exige escrita porque consome uma sessão/licença no ERP a cada chamada.",
        "operationId": "protheus_testar_api_protheus_test_connection_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtheusTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/status": {
      "get": {
        "tags": [
          "banco"
        ],
        "summary": "Database Status",
        "description": "Diagnóstico da conexão do PORTAL, em leitura.\n\nA conexão é definida no ambiente, não pela tela — por isso não existe rota\nde escrita aqui. A senha nunca aparece, nem mascarada.",
        "operationId": "database_status_api_database_status_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostics": {
      "get": {
        "tags": [
          "diagnostico"
        ],
        "summary": "Diagnostico",
        "operationId": "diagnostico_api_diagnostics_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "tp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaminhoDiagnostico": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "caminho": {
            "type": "string",
            "title": "Caminho"
          },
          "existe": {
            "type": "boolean",
            "title": "Existe"
          },
          "tamanho": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tamanho"
          },
          "observacao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacao"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "caminho",
          "existe"
        ],
        "title": "CaminhoDiagnostico"
      },
...
      "DatabaseStatus": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "servidor": {
            "type": "string",
            "title": "Servidor"
          },
          "porta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Porta"
          },
          "banco": {
            "type": "string",
            "title": "Banco"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "criptografado": {
            "type": "boolean",
            "title": "Criptografado"
          },
          "conectado": {
            "type": "boolean",
            "title": "Conectado"
          },
          "latenciaMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latenciams"
          },
          "versaoServidor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Versaoservidor"
          },
          "detalheErro": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detalheerro"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "servidor",
          "porta",
          "banco",
          "usuario",
          "criptografado",
          "conectado",
          "latenciaMs",
          "versaoServidor",
          "detalheErro"
        ],
        "title": "DatabaseStatus",
        "description": "Diagnóstico da conexão do PORTAL, em leitura.\n\nA conexão é definida no ambiente, não pela tela. A senha nunca aparece\naqui, nem mascarada."
      },
      "DiagnosticoResponse": {
        "properties": {
          "geradoEm": {
            "type": "string",
            "format": "date-time",
            "title": "Geradoem"
          },
          "runtime": {
            "items": {
              "$ref": "#/components/schemas/ItemDiagnostico"
            },
            "type": "array",
            "title": "Runtime"
          },
          "caminhos": {
            "items": {
              "$ref": "#/components/schemas/CaminhoDiagnostico"
            },
            "type": "array",
            "title": "Caminhos"
          },
          "pacotes": {
            "items": {
              "$ref": "#/components/schemas/PacoteDiagnostico"
            },
            "type": "array",
            "title": "Pacotes"
          },
          "banco": {
            "items": {
              "$ref": "#/components/schemas/ItemDiagnostico"
            },
            "type": "array",
            "title": "Banco"
          },
          "alertas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alertas"
          }
        },
        "type": "object",
        "required": [
          "geradoEm",
          "runtime",
          "caminhos",
          "pacotes",
          "banco",
          "alertas"
        ],
        "title": "DiagnosticoResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemDiagnostico": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "valor": {
            "type": "string",
            "title": "Valor"
          },
          "observacao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacao"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "valor"
        ],
        "title": "ItemDiagnostico"
      },
...
      "PacoteDiagnostico": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "versao": {
            "type": "string",
            "title": "Versao"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "versao"
        ],
        "title": "PacoteDiagnostico"
      },
...
      "ProtheusStatus": {
        "properties": {
          "configurado": {
            "type": "boolean",
            "title": "Configurado"
          },
          "baseUrl": {
            "type": "string",
            "title": "Baseurl"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "timeoutSegundos": {
            "type": "integer",
            "title": "Timeoutsegundos"
          },
          "senhaConfigurada": {
            "type": "boolean",
            "title": "Senhaconfigurada"
          }
        },
        "type": "object",
        "required": [
          "configurado",
          "baseUrl",
          "usuario",
          "timeoutSegundos",
          "senhaConfigurada"
        ],
        "title": "ProtheusStatus",
        "description": "O que está configurado. A senha nunca aparece, só se ela existe."
      },
      "ProtheusTestResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "detalhe": {
            "type": "string",
            "title": "Detalhe"
          },
          "tokenValidoPorSegundos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokenvalidoporsegundos"
          },
          "duracaoMs": {
            "type": "integer",
            "title": "Duracaoms"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "detalhe",
          "tokenValidoPorSegundos",
          "duracaoMs"
        ],
        "title": "ProtheusTestResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}